+++ packages/af-utils-react-virtual-headless/src/hooks/useVirtualModel/index.js
@@ -1,25 +1,15 @@
 import List from "models/List";
-import {
-    DEFAULT_OVERSCAN_COUNT,
-    DEFAULT_ESTIMATED_WIDGET_SIZE,
-    DEFAULT_ESTIMATED_ITEM_SIZE
-} from "constants";
+import { DEFAULT_ESTIMATED_WIDGET_SIZE } from "constants";
+import mergeModelParams from "/utils/mergeModelParams";
 
 import useOnce from "../useOnce";
 
-const useVirtualModel = ({
-    itemCount = 0,
-    estimatedItemSize = DEFAULT_ESTIMATED_ITEM_SIZE,
-    estimatedWidgetSize = DEFAULT_ESTIMATED_WIDGET_SIZE,
-    overscanCount = DEFAULT_OVERSCAN_COUNT,
-    horizontal
-}) =>
+const useVirtualModel = params =>
     useOnce(() => {
-        const model = new List(estimatedWidgetSize);
-
-        model.setOverscan(overscanCount);
-        model.setHorizontal(!!horizontal);
-        model.setItemCount(itemCount, estimatedItemSize);
+        const model = new List(
+            params.estimatedWidgetSize ?? DEFAULT_ESTIMATED_WIDGET_SIZE
+        );
+        mergeModelParams(model, params);
 
         return model;
     });